refactor(Domain): 메뉴, 상품 객체 테스트 분리
